requirements.txt
bcrypt==4.1.3
mysql-connector-python==8.4.0
